feat(kombiner/v1alpha): move kombiner v1alpha under its own directory under apis to create a dir hierarchy for a new to-be-introduced config type
apis/
├── kombiner/
│   └── v1alpha1/
│       ├── groupversion_info.go
│       ├── types.go
│       └── zz_generated.deepcopy.go
├── config/
│   └── v1alpha1/
