+++ pkg/scheduler/bind_plugin_test.go
@@ -17,7 +17,7 @@ import (
 	"k8s.io/kubernetes/pkg/scheduler/framework"
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
 
-	"kombiner/pkg/apis/v1alpha1"
+	"kombiner/pkg/apis/kombiner/v1alpha1"
 	"kombiner/pkg/generated/clientset/versioned/fake"
 )
